Add 'global' driver option to erl_ddll:try_load/3
On systems using dlopen() to load a driver, it may be necessary
to use the RTLD_GLOBAL flag.

This patch adds support for the 'global' atom in 'driver_options'
Using {driver_options, [global]} will set the RTLD_FLAG.

Parts of erts using dlopen() unrelated to try_load/3 have been
modified to use only RTLD_NOW, as it is the current behaviour.

This option will be ignored on platforms without dlopen().